requirements.txt
hyperbrowser==0.79.0
openai==2.14.0
playwright==1.57.0
python-dotenv==1.2.1
streamlit==1.52.2
